Removed additional files, fixed scrolling
